Create README.md
This project is scraped from Google Play Store Apps and is available on Kaggle.com. This project can help app-making businesses to success. This project provides insights to mobile app developers in terms of which kind of apps should they develop to gain maximum revenue. It also trie and predicts the rating of an app. Each app (row) has values for catergory, rating, size, and more. In this project we have done EDA studies as well as predict ratings for apps.